Refactor: Componentize Homepage.tsx for improved readability and maintainability
- Extracted major sections of the Homepage into individual components:
  - HeroSection (including search bar and suggestions)
  - TrustIndicators
  - WriteReviewPrompt
  - BestEntitiesSection
  - HowItWorksSection

- Moved data fetching and processing logic (for best entities and recent reviews) into a custom hook `useHomepageData`.

- Relocated helper functions (getUserInitials, getAvatarColor) to `src/utils/uiHelpers.ts`.

- Simplified the main `Homepage.tsx` component to primarily compose these new components and utilize the data hook.

This refactoring enhances the structure of the homepage, making it easier to understand, manage, and extend.
